Update _global.scss
updated the padding for the one-pager to eliminate the extra white space. 
I think it can be removed completely as the grid seems to be good with the spacing in both mobile and web view... but for now I kept it on the "left" just in case we do something with the text.